Move kubernetes Sync code into sync.go
Just a rearrangement, no logic changes.
